% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_cp949.R
\name{bit_cp949}
\alias{bit_cp949}
\title{Convert Encoding of Dataframe from CP949 to UTF-8}
\usage{
bit_cp949()
}
\value{
This function returns no value. It alters the chosen dataframe in the global environment in-place.
}
\description{
This function changes the encoding of the character columns in a dataframe from CP949 to UTF-8.
}
\details{
The function provides a GUI for the user to select a dataframe from the global environment. It then changes the encoding of character columns and column names of the chosen dataframe from CP949 to UTF-8.
}
\note{
This function requires the gWidgets2 and data.table packages.
}
\examples{
\dontrun{
 # Generate a sample dataframe
 df <- data.frame(a = c("가", "나"), b = c("다", "라"))
 Encoding(df$a) <- Encoding(df$b) <- "CP949"
 assign("df", df, envir = .GlobalEnv)
 # Run the function
 bit_cp949()
}
}
\seealso{
\code{\link[data.table]{data.table}}, \code{\link[gWidgets2]{gwindow}}
}
