% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_csv.R
\name{bit_csv}
\alias{bit_csv}
\title{Load a CSV file using a dialog}
\usage{
bit_csv(file = NULL)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file to load.
If NULL (the default), a file dialog will be shown.}
}
\value{
NULL. This function is called for its side effect of loading a data frame into the global environment.
}
\description{
This function allows the user to choose a CSV file from their file system and
then loads that file into a data frame in the global environment.
}
