% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_csv_all.R
\name{bit_csv_all}
\alias{bit_csv_all}
\title{Load all CSV files in a directory}
\usage{
bit_csv_all()
}
\value{
NULL. This function is called for its side effect of loading data frames into the global environment.
}
\description{
This function allows the user to choose a directory and loads all .csv files
in that directory into separate data frames in the global environment.
}
