% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_editor.R
\name{bit_editor}
\alias{bit_editor}
\title{bit_editor: A simple text editor with syntax highlighting and scripting capabilities}
\usage{
bit_editor()
}
\value{
A tktoplevel widget that acts as a text editor.
}
\description{
This function creates a Tcl/Tk text editor with basic syntax highlighting,
scripting capabilities, load/save functionality, and a scrollbar.
}
\details{
This function utilizes several nested functions to achieve its functionality:
1. `save`: Saves the current content of the text box to a file.
2. `load`: Loads the content of a file into the text box.
3. `run`: Executes the current content of the text box as R code.

The text editor supports basic syntax highlighting. It highlights reserved words
in R programming with a blue color. It also has a scrollbar to navigate through
large amounts of text and supports text wrapping.

The editor also features a menu bar with options to load a file, save the current
content to a file, and execute the current content as R code.
}
\note{
The syntax highlighting feature is quite rudimentary and highlights
reserved words even if they are part of comments or string literals.
}
\examples{
\dontrun{
bit_editor()
}

}
\seealso{
\code{\link{tktext}}, \code{\link{tkscrollbar}}, \code{\link{tkmenu}}
}
