% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bit_setwd.R
\name{bit_setwd}
\alias{bit_setwd}
\title{Set Working Directory Interactively}
\usage{
bit_setwd()
}
\value{
This function returns no value. It changes the working directory to the user-selected directory in-place.
}
\description{
This function provides a GUI for selecting a working directory.
}
\details{
The function uses tcltk::tkchooseDirectory() to let the user choose a directory from their file system. After a directory is chosen, the function changes the working directory to the chosen directory and prints a message with the new directory and the corresponding setwd() code.
}
\note{
This function requires the tcltk package.
}
\examples{
\dontrun{
 # Run the function
 bit_setwd()
}
}
\seealso{
\code{\link[base]{setwd}}
}
